Convert eof of 5 *.cs files from CRLF to LF in rep
Generally all text files in the repo are normalized to LF line ending. The
five files in this commit had CRLF in the repo, which they should not.
This must have happened when someone commited them with autocrlf=false
while .gitattributes was not yet present. This commit only fixes the eof
of these five files in the repo, so that from now on all files in the repo
have LF line ending again.