Changed default Graph layout method and moved default Sugiyama code
The Graph is always set to CompoundFDP to start, so that single-commit
repos do not crash the graph.  The if statement is moved to after the
initial Graph = _graphBuilder.Graph() to prevent a crash.  When
switching from a single-commit repo to a many-commit repo, SeeGit will
crash if the method is set to EfficientSugiyama in Graph()