Rename `master` to `main`
The Internet Engineering Task Force (IEFT) points out that "Master-slave is an oppressive metaphor (https://tools.ietf.org/id/draft-knodel-terminology-00.html#rfc.section.1.1.1) that will and should never become fully detached from history" as well as "In addition to being inappropriate and arcane, the master-slave metaphor is both technically and historically inaccurate."

It also turns out that the `master` branch naming of Git is not from "master" copy but is a legacy of its BitKeeper origins and master/slave repositories.
https://mail.gnome.org/archives/desktop-devel-list/2019-May/msg00066.html

This renames the `master` branch to `main` which is a more accurate description and uses more inclusive language.